@@ -7,22 +7,17 @@ import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.kira.mypublishplatform.R
 import com.kira.mypublishplatform.base.BaseActivity
+import com.kira.mypublishplatform.databinding.ActivityMainBinding
 import kotlinx.android.synthetic.main.activity_main.*
 
 
-class MainActivity : BaseActivity() {
+class MainActivity : BaseActivity<ActivityMainBinding>() {
 
-//
-
-    override fun loadXml() {
-        setContentView(R.layout.activity_main)
-    }
+//    override fun loadXml() {
+//        setContentView(R.layout.activity_main)
+//    }
 
     override fun getIntentData(savedInstanceState: Bundle?) {
-
-    }
-
-    override fun initView() {
 
     }
 
@@ -32,7 +27,7 @@ class MainActivity : BaseActivity() {
             setOf(
                 R.id.navigation_home,
                 R.id.navigation_dashboard,
-                R.id.navigation_notifications
+                R.id.navigation_mine
             )
         )
         val navController = findNavController(R.id.nav_host_fragment)
