@@ -1,12 +1,40 @@
 package com.kira.mypublishplatform.model
 
-import java.io.Serializable
+import android.os.Parcel
+import android.os.Parcelable
 
-class SubModel : Serializable {
-    var name: String? = null
-    var icon: String? = null
-    var href: String? = null
-    var enable: Boolean? = null
-    var needLogin: Boolean? = null
+
+data class SubModel(
+    var name: String?, var icon: String?, var href: String?, var enable: Boolean?, var needLogin: Boolean?
+) : Parcelable {
+    constructor(parcel: Parcel) : this(
+        parcel.readString(),
+        parcel.readString(),
+        parcel.readString(),
+        parcel.readValue(Boolean::class.java.classLoader) as? Boolean,
+        parcel.readValue(Boolean::class.java.classLoader) as? Boolean
+    )
+
+    override fun writeToParcel(parcel: Parcel, flags: Int) {
+        parcel.writeString(name)
+        parcel.writeString(icon)
+        parcel.writeString(href)
+        parcel.writeValue(enable)
+        parcel.writeValue(needLogin)
+    }
+
+    override fun describeContents(): Int {
+        return 0
+    }
+
+    companion object CREATOR : Parcelable.Creator<SubModel> {
+        override fun createFromParcel(parcel: Parcel): SubModel {
+            return SubModel(parcel)
+        }
+
+        override fun newArray(size: Int): Array<SubModel?> {
+            return arrayOfNulls(size)
+        }
+    }
 
 }
